Fix issues caused by accessing destroyed entities

+++ samples/asteroids/components/projectile.rb
@@ -4,11 +4,11 @@ module Asteroids
   class Projectile < Engine::Component
     def update(delta_time)
       AsteroidComponent.asteroids.each do |asteroid|
+        next if asteroid.destroyed?
         if (game_object.pos - asteroid.game_object.pos).magnitude < asteroid.size
-          game_object.destroy!
           asteroid.blow_up
-
           Explosion.create(game_object.pos)
+          game_object.destroy!
           break
         end
       end
